Fix the client error

+++ cleaner.rb
@@ -1,34 +1,48 @@
 require 'aws-sdk'
 require 'awesome_print'
+require 'time'
 
 akid = ENV.fetch "AWS_ACCESS_KEY_ID"
 secret = ENV.fetch "AWS_SECRET_ACCESS_KEY"
 region = ENV.fetch "AWS_DEFAULT_REGION"
+age = ENV.fetch "AGE_DAYS"
+
+clean_time = Time.now - age.to_i * 86400
 
 Aws.config.update({
   region: region,
   credentials: Aws::Credentials.new(akid, secret),
 })
 
-beanstalk = Aws::ElasticBeanstalk::Client.new
+client = Aws::ElasticBeanstalk::Client.new
 
-resp = beanstalk.describe_applications
+resp = client.describe_applications
 resp.applications.each do |app|
-  app.versions.each do |version|
-    env = beanstalk.describe_environments({
+    versions_result = client.describe_application_versions({
       application_name: app.application_name,
-      version_label: version,
+      version_labels: app.versions
     })
 
-    unless env.environments.any?
-      ap "There are no apps with version #{version}"
-      client.delete_application_version({
+  versions_result.application_versions.each do |version|
+    if version.date_created < clean_time
+      env = client.describe_environments({
         application_name: app.application_name,
-        version_label: version,
-        delete_source_bundle: true,
+        version_label: version.version_label
       })
+
+      unless env.environments.any?
+        ap "There are no apps with version #{version}, deleting..."
+        client.delete_application_version({
+          application_name: app.application_name,
+          version_label: version,
+          delete_source_bundle: true
+        })
+      end
+    else
+      ap "The app (#{app.application_name}) version (#{version.version_label}) is not old enough to delete (#{version.date_created} > #{clean_time})"
     end
 
-    sleep 1.5
+    # Sleep to prevent ratelimit
+    sleep 1
   end
 end
